+++ src/execution/index.ts
@@ -1,6 +1,9 @@
-export * from './iexecute_process_service';
 export * from './execute_process_service';
-export * from './next_flow_node_info';
 export * from './handler';
+export * from './iexecute_process_service';
+export * from './iprocess_model_fascade';
+export * from './iprocess_token_fascade';
+export * from './iprocess_token_result';
+export * from './next_flow_node_info';
 export * from './process_model_fascade';
 export * from './process_token_fascade';
