+++ src/execution/iprocess_token_result.ts
@@ -1,0 +1,5 @@
+
+export interface IProcessTokenResult {
+  flowNodeId: string;
+  result: any;
+}
